Update backpacks.ts
nerf to level 2 and 3 backpack ammo/grenade capacity
in my opinion, the backpack capacity scaling is way off, since all necesity of resource management goes out the window the moment you get a regular or tactical backpack. This change makes the tactical backpack behave more like the regular one, and moves the regular backpack to somewhere closer to the level 1.