refactor: improve packet api usability
the big idea of the packet api is that each packet possess both the means to write/read itself from a stream (and thus, facilities for containing and exposing relevant data); this has been kept, but the means by which a packet is made is more structured now. notably, packets should now be type-safe for external users. (killfeed packet remains somewhat finicky to create, but should be better)
this is not meant to be a behaviorial change.

additionally, and on a lighter note, stress test bots can now randomly aimlock onto another bot.