perf: create faster versions of input for reuse
old system worked by associating each key with a string, and then having the console parse that string when the input is invoked

new system also does this on the first try, but the console now "compiles" the query into a much faster-to-execute js function that is reused for future invocations of that input.

test was run measuring the time it took for the game to find and dispatch all actions related to a given input (specifically, the entirety of the `InputManager#handleInputEvent` method)
the new system is about 40–50% faster.