fix: big teammate bug (hopefully) (read desc)
When a player initially connects, they're spawned in and added to a team, but not fully activated until a JoinPacket is received. If that JoinPacket was never received for whatever reason, it might cause ghost teammates and similar issues.
With this change, players are automatically disconnected if they don't send a JoinPacket within 5 seconds of connecting.