packet api v5
this newest iteration of the packet api sees the move away from bit-wise operations towards faster byte-wise ones. this means the removal of the bit-buffer dependancy, in favor of the perhaps uncreatively named 'byte stream', which acts as a wrapper around a DataView
in isolated, this 'byte stream' can perform about twice as fast as the previous bit-buffer
ingame, the speed gain is not as great, but cpu profiles indicate that it is not negligable either.